Refactor grudge.op and modularize
